requirements.txt
pyTelegramBotAPI==4.20.0
python-dotenv==1.0.0
